… src/Ambev.DeveloperEvaluation.Application/Sales/RemoveSaleItem/RemoveSaleItemCommandValidator.cs
… src/Ambev.DeveloperEvaluation.Application/Sales/RemoveSaleItem/RemoveSaleItemHandler.cs
… src/Ambev.DeveloperEvaluation.Application/Sales/RemoveSaleItem/RemoveSaleItemResult.cs
… src/Ambev.DeveloperEvaluation.Application/Sales/UpdateSaleItem/UpdateSaleItemCommand.cs
… src/Ambev.DeveloperEvaluation.Application/Sales/UpdateSaleItem/UpdateSaleItemCommandValidator.cs
… src/Ambev.DeveloperEvaluation.Application/Sales/UpdateSaleItem/UpdateSaleItemHandler.cs
… src/Ambev.DeveloperEvaluation.Application/Sales/UpdateSaleItem/UpdateSaleItemResult.cs
… src/Ambev.DeveloperEvaluation.Domain/Entities/Sale.cs
… src/Ambev.DeveloperEvaluation.Domain/Entities/SaleItem.cs
… src/Ambev.DeveloperEvaluation.Domain/Services/ISaleService.cs
… src/Ambev.DeveloperEvaluation.IoC/ModuleInitializers/InfrastructureModuleInitializer.cs
… tests/Ambev.DeveloperEvaluation.Unit/Application/Sales/AddSaleItem/AddSaleItemCommandValidatorTests.cs
… tests/Ambev.DeveloperEvaluation.Unit/Application/Sales/AddSaleItem/AddSaleItemHandlerTests.cs
… tests/Ambev.DeveloperEvaluation.Unit/Application/Sales/RemoveSaleItem/RemoveSaleItemHandlerTests.cs
… tests/Ambev.DeveloperEvaluation.Unit/Application/Sales/UpdateSaleItem/UpdateSaleItemHandlerTests.cs
… tests/Ambev.DeveloperEvaluation.Unit/Common/TestData/ProductTestData.cs
… tests/Ambev.DeveloperEvaluation.Unit/Domain/Entities/SaleItemTests.cs
… tests/Ambev.DeveloperEvaluation.Unit/Domain/Entities/SaleTests.cs