This pull request introduces a comprehensive set of features for managing sale items within the application, including commands, validators, handlers, and result models for adding, removing, and updating sale items. It also strengthens business rule enforcement in the domain layer and adjusts a timing constant for validation. The changes collectively enhance the application's ability to maintain data integrity and enforce business constraints around sales operations.

**New Application Layer Features**

* Added command, validator, handler, and result model for adding a sale item, including validation for sale/product existence, quantity limits, and optional unit price. (`AddSaleItemCommand`, `AddSaleItemCommandValidator`, `AddSaleItemHandler`, `AddSaleItemResult`) [[1]](diffhunk://#diff-15b4b87627672ed33a6c6c6c1555b39f2d31f23cebdee5b937b68011a7b700c8R1-R14) [[2]](diffhunk://#diff-af630a5575684826e32cf8a4e0d6de64967d16b6a35661aab1a1d478dd55bf6fR1-R31) [[3]](diffhunk://#diff-97539e2eb61013e637a504cd847f91eee03393686878eec2411942702da9ecbaR1-R72) [[4]](diffhunk://#diff-2a52f602901a7fe1058ada439b368d1789561812cc54fb546cfef5006de2721dR1-R16)
* Added command, validator, handler, and result model for removing a sale item, with validation and logic to restore product stock and mark items as cancelled rather than deleted. (`RemoveSaleItemCommand`, `RemoveSaleItemCommandValidator`, `RemoveSaleItemHandler`, `RemoveSaleItemResult`) [[1]](diffhunk://#diff-2a05c8779ee778c721af1c882b96abab91c568914a67b5d3b54094a14f8b9ab0R1-R12) [[2]](diffhunk://#diff-b50f6f07cbaaa3ff7f2b59732018f43c141cccb770b9b88bf357f768d23bc67fR1-R36) [[3]](diffhunk://#diff-91370069cfaa73ec74ab695405acfd9c2fd93d2d9612027f157ff21ccc6ca1c2R1-R79) [[4]](diffhunk://#diff-c80bc19218d67b8b7f8e63f3a55c624b5d32b4a65171da237941d99c7bdc8382R1-R53)
* Added command, validator, handler, and result model for updating the quantity of an existing sale item, including quantity difference calculation and appropriate stock adjustments. (`UpdateSaleItemCommand`, `UpdateSaleItemCommandValidator`, `UpdateSaleItemHandler`, `UpdateSaleItemResult`) [[1]](diffhunk://#diff-b819dd5312f033b4c54a89ab0debf8c11813888a1c648a0381eac282449c9b6cR1-R13) [[2]](diffhunk://#diff-3dceeab7b3fa5f2cec6b91e15e9948b1ba5a9bf1e615a0e930f9c099c97bc622R1-R26) [[3]](diffhunk://#diff-faee9ff8f7d27cd4a305f442156bcad7f22b4a610c06bb1f7d35c7dfb721be36R1-R94) [[4]](diffhunk://#diff-def0fcf5996ab45fd4d6427fbc6f28755fda20d9666212b08385f1c43e129f4bR1-R18)

**Domain Layer Business Rule Enhancements**

* Strengthened business rules in `Sale.AddItem` to check product stock availability and enforce a maximum of 20 different products per sale, in addition to the existing per-product quantity limit.
* Improved `Sale.UpdateItemQuantity` to validate stock availability when increasing quantity, ensuring sufficient product stock before allowing updates.

**Configuration Adjustment**

* Increased the date validation buffer from 100ms to 2000ms in `BusinessConstants` to better handle timing differences during entity validation.

## Summary by Sourcery

Implement a full CRUD workflow for sale items in the application layer and reinforce domain-level business rules around stock and product limits while expanding unit test coverage for these operations

New Features:
- Introduce AddSaleItem, RemoveSaleItem, and UpdateSaleItem application commands with their validators, handlers, and result models

Enhancements:
- Strengthen Sale domain rules to enforce stock availability, max 20 different products per sale, and validate stock on quantity updates
- Delegate basic quantity and price checks from SaleItem entity to application layer validators
- Increase entity date validation buffer from 100ms to 2000ms

Tests:
- Add comprehensive unit tests for application handlers and validators covering add, remove, and update sale item operations
- Adjust test data to increase product StockQuantity range and expand CreatedAt tolerance